+++ packages/auth0-auth-js/src/index.ts
@@ -1,3 +1,3 @@
-export { AuthClient } from './auth-client.js';
+export { AuthClient, SUBJECT_TYPE_ACCESS_TOKEN, SUBJECT_TYPE_REFRESH_TOKEN } from './auth-client.js';
 export * from './errors.js';
 export * from './types.js';